moved it into a component file and fixed conflicts
